Comments

+++ Backend/src/routes/roles.js
@@ -3,6 +3,7 @@ import db from "../config/db.js";
 
 const router = express.Router();
 
+// Endpoint to get all roles
 router.get("/", async (req, res) => {
   try {
     const [rows] = await db.query("SELECT * FROM roles ORDER BY id ASC");
@@ -13,6 +14,7 @@ router.get("/", async (req, res) => {
   }
 });
 
+// Endpoint to create a role
 router.get("/:roleId/permissions", async (req, res) => {
   const { roleId } = req.params;
 
@@ -22,10 +24,7 @@ router.get("/:roleId/permissions", async (req, res) => {
       [roleId]
     );
 
-    res.json({
-      ok: true,
-      permissions: rows.map(r => r.permission_id)
-    });
+    res.json({ ok: true, permissions: rows.map(r => r.permission_id)});
 
   } catch (error) {
     console.error("Error al obtener permisos del rol:", error);
@@ -33,6 +32,7 @@ router.get("/:roleId/permissions", async (req, res) => {
   }
 });
 
+// Endpoint to create a role
 router.post("/:roleId/permissions", async (req, res) => {
   const { roleId } = req.params;
   const { permissions } = req.body;
@@ -47,10 +47,7 @@ router.post("/:roleId/permissions", async (req, res) => {
     connection = await db.getConnection();
     await connection.beginTransaction();
 
-    await connection.query(
-      "DELETE FROM roles_permissions WHERE role_id = ?",
-      [roleId]
-    );
+    await connection.query( "DELETE FROM roles_permissions WHERE role_id = ?", [roleId]);
 
     for (const perm of permissions) {
       await connection.query(
@@ -73,6 +70,7 @@ router.post("/:roleId/permissions", async (req, res) => {
   }
 });
 
+// Endpoint to delete a role
 router.put("/update-role", async (req, res) => {
   let connection;
 
@@ -122,11 +120,7 @@ router.put("/update-role", async (req, res) => {
     if (newRoleId === 1)
       await connection.query("INSERT INTO admins (user_id) VALUES (?)", [userId]);
 
-    await connection.query(
-      "UPDATE users SET role_id = ? WHERE id = ?",
-      [newRoleId, userId]
-    );
-
+    await connection.query( "UPDATE users SET role_id = ? WHERE id = ?", [newRoleId, userId]);
     await connection.commit();
 
     res.json({ ok: true, message: "Rol actualizado exitosamente" });
@@ -135,7 +129,6 @@ router.put("/update-role", async (req, res) => {
     console.error(error);
 
     if (connection) await connection.rollback();
-
     res.status(500).json({ ok: false, message: "Error al actualizar rol" });
 
   } finally {
